+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.0] - 2021-03-30
+### Added
+- `SplashScreen` and `SimpleSplashScreen` widgets
+
+### Changed
+- `StyledText` now supports a `align` parameter (equivalent to `textAlign`)
+
+### Fixed
+- Dummy function in public API
+
+
 ## [1.2.0] - 2021-03-29
 ### Added
 - `SimplePadding` widget